Make pytest import django settings and fixtures only when needed.
